feat: add support for a user's package.json
This commit modifies the JS language implementation so that
it will look for a user provided package.json in the same directory
as the function file. If one exists, it uses this instead of the
generated one.

NB: This does mean that if a user has provided function dependency
information in the function file in the form:

```js
// kfn:dependency primal 0.2.3
```
These dependencies will be superceded by what is found in package.json