src: use the oscf/js-runtime image as the base
This commit changes the base image used for the Node.js runtime to use
`oscf/js-runtime` which already has a tagged and released version of
the `faas-js-runtime` module installed on it, and includes a simple
bootstrap shim to load the user function.

As a result of this change, there is no need to download the Node.js
runtime module directly, and `npm install` only needs to be run in
the user function directory.